tl_sync/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
mypy-extensions==0.4.3
psycopg2-binary==2.9.5
requests==2.28.1
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
